feat: Revert to seperate files
## Description
The previous changes merged the lib files into a single file due to issues
with switch to `@use` and `@forward`. With a bit more work on this it's
no longer a problem and having mutiple smaller files feels like it would
be easier to manage, so this update splits it back out again.

## Issue
https://jira.dev.bbc.co.uk/browse/GEL